Update to Comics
Comics have a series attribute to act as a category. Eventually this
could go into full blown category system but plain text for now. Comics
can now have a single attachment image that acts as a preview. Comics
also have a status attribute that defaults to `:regular` but may also be
`:featured`. The name `status` may be changed to better reflect that
this is more of a modification on the comic rather than it's actual
status, e.g. `Status: published, Mods: featured`. Time will tell. For
now good enough, ship it.